+++ share/completions/systemctl.fish
@@ -1,3 +1,8 @@
+
+# magic completion safety check (do not remove this comment)
+if not type -q systemctl
+    exit
+end
 set -l systemd_version (systemctl --version | string match "systemd*" | string replace -r "\D*(\d+)"  '$1')
 set -l commands list-units list-sockets start stop reload restart try-restart reload-or-restart reload-or-try-restart \
 isolate kill is-active is-failed status show get-cgroup-attr set-cgroup-attr unset-cgroup-attr set-cgroup help \
